refactor OrderEntity and MarketEntity

+++ core/domain/src/main/java/org/the_chance/honeymart/domain/model/MarketEntity.kt
@@ -2,7 +2,7 @@ package org.the_chance.honeymart.domain.model
 
 
 data class MarketEntity(
-    val marketId: Int?,
+    val marketId: Long?,
     val marketName: String?,
     val imageUrl: String?
 )
